src/discord.py
import os
import discord
from discord.ext import commands, tasks
import logging
from typing import Dict, List, Optional, Any, Union
from datetime import datetime
import pandas as pd
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Configure Discord
DISCORD_TOKEN = os.getenv("DISCORD_TOKEN")
DISCORD_CHANNEL_ID = int(os.getenv("DISCORD_CHANNEL_ID", "0"))

# Initialize Discord client
intents = discord.Intents.default()
intents.message_content = True
bot = commands.Bot(command_prefix="!", intents=intents)


@bot.event
async def on_ready() -> None:
    """Called when the bot is ready and connected to Discord."""
    logger.info(f"Bot is ready! Logged in as {bot.user}")
    send_trading_updates.start()


@tasks.loop(minutes=5)
async def send_trading_updates() -> None:
    """Send trading updates to Discord channel."""
    try:
        channel = bot.get_channel(DISCORD_CHANNEL_ID)
        if not channel:
            logger.error(f"Could not find channel with ID {DISCORD_CHANNEL_ID}")
            return

        # Get latest trading data
        df = get_latest_data()  # You'll need to implement this
        if df is None or df.empty:
            logger.warning("No data available for trading update")
            return

        # Get LLM analysis
        analysis = get_llm_response(df)  # You'll need to implement this
        if not analysis:
            logger.warning("No LLM analysis available")
            return

        # Format message
        message = format_trading_message(df, analysis)  # You'll need to implement this
        if not message:
            logger.warning("Failed to format trading message")
            return

        # Send message
        await channel.send(message)

    except Exception as e:
        logger.error(f"Error sending trading update: {str(e)}")


def format_trading_message(df: pd.DataFrame, analysis: Dict[str, Any]) -> Optional[str]:
    """
    Format trading message for Discord.

    Args:
        df: DataFrame with latest price data
        analysis: Dictionary containing LLM analysis

    Returns:
        Formatted message string or None if formatting fails
    """
    try:
        current_price = df["close"].iloc[-1]
        price_change = df["returns"].iloc[-1] * 100

        message = [
            "📊 Trading Update",
            "=" * 40,
            f"Current Price: ${current_price:.2f} ({price_change:+.2f}%)",
            "",
            "Analysis:",
            f"Recommendation: {analysis['recommendation']}",
            f"Position Size: {analysis['position_size']*100:.1f}%",
            f"Trade Type: {analysis['trade_type']}",
            "",
            "Price Targets:",
            f"Stop Loss: ${analysis['price_targets']['stop_loss']:.2f}",
            f"Take Profit: ${analysis['price_targets']['take_profit']:.2f}",
            "",
            "Technical Indicators:",
            f"RSI: {df['RSI'].iloc[-1]:.2f}",
            f"MACD: {df['MACD'].iloc[-1]:.2f}",
            f"Volume Ratio: {df['volume_ratio'].iloc[-1]:.2f}",
            "",
            "Market Context:",
            f"Trend: {analysis['market_context']['trend']}",
            f"Volatility: {analysis['market_context']['volatility']}",
            f"Volume: {analysis['market_context']['volume']}",
        ]

        return "\n".join(message)

    except Exception as e:
        logger.error(f"Error formatting trading message: {str(e)}")
        return None


def get_latest_data() -> Optional[pd.DataFrame]:
    """
    Get latest trading data.
    You'll need to implement this based on your data source.
    """
    # TODO: Implement data fetching
    return None


def get_llm_response(df: pd.DataFrame) -> Optional[Dict[str, Any]]:
    """
    Get LLM analysis for the latest data.
    You'll need to implement this based on your LLM integration.
    """
    # TODO: Implement LLM analysis
    return None


# Start the bot
if __name__ == "__main__":
    if not DISCORD_TOKEN:
        logger.error("Discord token not found in environment variables")
    elif not DISCORD_CHANNEL_ID:
        logger.error("Discord channel ID not found in environment variables")
    else:
        bot.run(DISCORD_TOKEN)
