requirements.txt
alpaca-py>=0.8.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
discord.py>=2.3.0
requests>=2.31.0
pytest>=7.0.0